+++ src/components/CodeFileName/CodeFileName.scss
@@ -1,4 +1,6 @@
-.filename {
+@import '../../styles/variables';
+
+.CodeFileName {
   font-size: $font-size - 0.2;
   background: #555;
   padding: 0.6rem;
