+++ src/components/CodeFileName/CodeFileName.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './CodeFileName.scss';
+
+export const CodeFileName: React.FC = ({ children }) => (
+  <div className="CodeFileName">{children}</div>
+);
