+++ src/components/CodeFileName/index.tsx
@@ -1,0 +1,1 @@
+export * from './CodeFileName';
