+++ src/components/Link/Link.scss
@@ -1,0 +1,36 @@
+.Link--button {
+  color: var(--text-color);
+  background-color: transparent;
+  border: none;
+  padding: 0.4em 0.6rem;
+  cursor: pointer;
+  text-align: center;
+  justify-content: center;
+  white-space: nowrap;
+  border-radius: 0.35rem;
+  text-decoration: none;
+  transition: background, background-color 0.35s ease;
+  outline: none;
+
+  /** Remove link styles */
+  &:after {
+    display: none;
+  }
+
+  &:visited {
+    color: var(--text-color);
+  }
+
+  &:focus {
+    box-shadow: none;
+  }
+
+  &.Link--button--active,
+  &:hover,
+  &:active,
+  &:focus {
+    text-decoration: none;
+    background-color: var(--link-hover-background);
+    color: var(--link-color);
+  }
+}
