+++ src/components/Navbar/Navbar.scss
@@ -1,4 +1,4 @@
-@import '../../styles/tools';
+@import '../../styles/mixins';
 @import '../../styles/utilities';
 
 .Navbar {
