+++ src/components/PostAuthor/PostAuthor.scss
@@ -1,4 +1,4 @@
-@import '../../styles/tools';
+@import '../../styles/mixins';
 
 .PostAuthor {
   display: flex;
