+++ src/components/ProfileHeader/ProfileHeader.scss
@@ -1,4 +1,4 @@
-@import '../../styles/tools';
+@import '../../styles/mixins';
 
 .ProfileHeader {
   display: flex;
