+++ src/styles/components/index.scss
@@ -1,4 +1,2 @@
-@import 'button';
 @import 'hero';
 @import 'toggle';
-@import 'filename';
