+++ src/styles/index.scss
@@ -1,6 +1,7 @@
-@import 'settings';
-@import 'tools';
+@import 'variables';
+@import 'mixins';
 @import 'generic';
 @import 'elements';
 @import 'components';
 @import 'utilities';
+
