--- src/styles/tools/index.scss
@@ -1,1 +1,0 @@
-@import 'mixins';
